healthphone/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
gunicorn==20.1.0

# AWS CLI y Elastic Beanstalk
awscli  # AWS Command Line Interface
boto3  # SDK de AWS para Python

# Para trabajar con Docker
#docker==5.0.3
#docker-compose==1.29.2

# Conector MySQL para Python
mysql-connector-python==9.0.0 # Conector MySQL para Django
mysqlclient==2.2.4

# Librerías adicionales
#requests # Para realizar solicitudes HTTP
python-decouple==3.5  # Para manejar variables de entorno
python-dateutil==2.8.2  # Manipulación de fechas
pytz==2023.3  # Manejo de zonas horarias
setuptools  # Asegurarse de que setuptools esté instalado
django-extensions
pydotplus
